requirements.txt
pandas
fastapi
python-dotenv
pydantic
pycaret
celery
flower
aiohttp
cerberus
uvicorn
